requirements.txt
streamlit>=1.33.0
openai>=1.12.0
azure-cognitiveservices-speech>=1.34.0
python-dotenv>=1.0.0